# openapi.json
{
  "components": {
    "schemas": {
      "ChatCompletionRequest": {
        "example": {
          "max_tokens": 64,
          "messages": [
            {
              "content": "Hello, Atlas.",
              "role": "user"
            }
          ],
          "model": "mock",
          "stream": false,
          "temperature": 0.2
        },
        "properties": {
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cap on generated tokens. `None` defers to the provider default.",
            "title": "Max Tokens"
          },
          "messages": {
            "description": "Ordered conversation messages.",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "model": {
            "description": "Provider model id or Atlas alias (resolved via `app.providers.registry`).",
            "title": "Model",
            "type": "string"
          },
          "prompt_params": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Template params for `prompt_ref` rendering.",
            "title": "Prompt Params"
          },
          "prompt_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional prompt-registry reference `<name>@<semver|production>` (REG-4). Resolved + rendered server-side and prepended as a system message. `None` (default) leaves the request unchanged.",
            "title": "Prompt Ref"
          },
          "stream": {
            "default": false,
            "description": "When `true`, response is `text/event-stream` carrying `chat.completion.chunk` deltas terminated by `data: [DONE]`.",
            "title": "Stream",
            "type": "boolean"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sampling temperature in `[0, 2]`. `None` defers to the provider default.",
            "title": "Temperature"
          }
        },
        "required": [
          "model",
          "messages"
        ],
        "title": "ChatCompletionRequest",
        "type": "object"
      },
      "ChatMessage": {
        "example": {
          "content": "Hello, Atlas.",
          "role": "user"
        },
        "properties": {
          "content": {
            "description": "Message content (plain text).",
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "Conversation role. One of `system`, `user`, `assistant`, `tool`.",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "Typed error body for non-2xx responses.\n\nFastAPI's default HTTPException emits `{\"detail\": ...}`. Some paths (rate\nlimit, budget) return a richer body — this envelope is the documented\nsuperset so codegen produces a single typed error model.",
        "example": {
          "code": "unauthorized",
          "detail": "invalid api key"
        },
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional machine-readable error code (provider/guardrail name).",
            "title": "Code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ],
            "description": "Human-readable error message, or a structured object for guardrail / rate-limit rejections (see per-endpoint response docs).",
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "Atlas API key",
        "description": "Atlas API key sent as `Authorization: Bearer <key>`. Local dev default is `dev-key`; real deployments source keys from Key Vault.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "contact@bragdev.com",
      "name": "Bruno Braga"
    },
    "description": "The **Atlas Gateway** is the single OpenAI-compatible facade in front of\nOpenAI, Anthropic, and Google. Every LLM call from the frontend and the agent\nruntime flows through here so prompts are versionable, guardrails enforceable,\nand every call cost-traceable (ADR-014).\n\n### Authentication\nAll `/v1/*` endpoints require `Authorization: Bearer <key>`. Local dev keys\ncome from `ATLAS_API_KEYS` (default `[\"dev-key\"]`); real deployments source\nkeys from Key Vault.\n\n### Streaming\n`POST /v1/chat/completions` with `\"stream\": true` returns\n`text/event-stream` carrying OpenAI `chat.completion.chunk` deltas terminated\nby `data: [DONE]`.\n\n### Errors\n4xx/5xx responses use the `ErrorEnvelope` schema (`{detail, code?}`).\nPer-endpoint codes are documented under each operation's responses.\n",
    "license": {
      "name": "Internal — Atlas (Enhesa mirror)"
    },
    "title": "Atlas Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "description": "OpenAI-compatible chat completion. Set `stream: true` to receive a `text/event-stream` of `chat.completion.chunk` deltas instead of a single JSON response. Honors guardrails (GRD-*), rate limits (GW-16), monthly budgets (GW-17), and the prompt registry (REG-4).",
        "operationId": "chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "example": "data: {\"id\":\"chatcmpl-1\",\"object\":\"chat.completion.chunk\",\"created\":1735689600,\"model\":\"mock\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\"}}]}\n\ndata: {\"id\":\"chatcmpl-1\",\"object\":\"chat.completion.chunk\",\"created\":1735689600,\"model\":\"mock\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\"Hi\"}}]}\n\ndata: [DONE]\n\n",
                "schema": {
                  "description": "Server-Sent Events. Each line `data: {json}` carries a `ChatCompletionChunk`. Terminator: `data: [DONE]`.",
                  "type": "string"
                }
              }
            },
            "description": "Chat completion (JSON) or SSE stream of `chat.completion.chunk` deltas."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid Bearer key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown model — no provider/alias matched."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Guardrail rejection. `detail` is `{guardrail, phase, reason}` naming the check and pipeline phase that rejected the request."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit or monthly budget exceeded. Response carries `Retry-After` (seconds). Body shape matches atlas-docs/03 §5.2."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a chat completion",
        "tags": [
          "chat"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local debug (VS Code launcher)",
      "url": "http://localhost:8090"
    },
    {
      "description": "In-cluster (AKS)",
      "url": "http://atlas-gateway.atlas-platform.svc:8000"
    }
  ],
...
}